# frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "园智汇 · AI 产业园区智慧运营管理平台",
  "description": "SmartPark AI Operations Platform\n\n面向产业园区集团的多租户智慧运营平台，包含数据驾驶舱、空间资产、企业服务、招商运营、合同财务、项目管理、物业运维、安全管理、审批中心、AI Agent（1 个主 Agent + 11 个子 Agent）、数据中心、报表中心、日志审计与系统管理。\n\n**数据说明**：演示环境，全部数据为演示数据（is_demo=True），由数据库实时聚合计算。",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/space/parks": {
   "get": {
    "tags": [
     "空间资产"
    ],
    "summary": "List Parks",
    "operationId": "list_parks_api_space_parks_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response List Parks Api Space Parks Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ]
   }
  },
  "/api/space/parks/{park_id}/twin": {
   "get": {
    "tags": [
     "空间资产"
    ],
    "summary": "Park Twin",
    "description": "2.5D 数字孪生数据：楼栋坐标 + 实时状态。",
    "operationId": "park_twin_api_space_parks__park_id__twin_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "park_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Park Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Park Twin Api Space Parks  Park Id  Twin Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/space/buildings": {
   "get": {
    "tags": [
     "空间资产"
    ],
    "summary": "List Buildings",
    "operationId": "list_buildings_api_space_buildings_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "park_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Park Id"
      }
     },
     {
      "name": "keyword",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Keyword"
      }
     },
     {
      "name": "building_type",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Building Type"
      }
     },
     {
      "name": "level",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Level"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 20,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Buildings Api Space Buildings Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/space/buildings/{building_id}": {
   "get": {
    "tags": [
     "空间资产"
    ],
    "summary": "Building Detail",
    "description": "一楼一档。",
    "operationId": "building_detail_api_space_buildings__building_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "building_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Building Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Building Detail Api Space Buildings  Building Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/space/spaces": {
   "get": {
    "tags": [
     "空间资产"
    ],
    "summary": "List Spaces",
    "operationId": "list_spaces_api_space_spaces_get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "park_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Park Id"
      }
     },
     {
      "name": "building_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Building Id"
      }
     },
     {
      "name": "floor_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Floor Id"
      }
     },
     {
      "name": "space_type",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Space Type"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     },
     {
      "name": "keyword",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Keyword"
      }
     },
     {
      "name": "min_area",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "Min Area"
      }
     },
     {
      "name": "max_area",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number"
        },
        {
         "type": "null"
        }
       ],
       "title": "Max Area"
      }
     },
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 1,
       "default": 1,
       "title": "Page"
      }
     },
     {
      "name": "page_size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 200,
       "minimum": 1,
       "default": 20,
       "title": "Page Size"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Spaces Api Space Spaces Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/space/spaces/{space_id}": {
   "get": {
    "tags": [
     "空间资产"
    ],
    "summary": "Space Detail",
    "operationId": "space_detail_api_space_spaces__space_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "space_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Space Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Space Detail Api Space Spaces  Space Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/space/spaces/board/{building_id}": {
   "get": {
    "tags": [
     "空间资产"
    ],
    "summary": "Space Board",
    "description": "一房一状态：楼栋楼层-空间矩阵视图。",
    "operationId": "space_board_api_space_spaces_board__building_id__get",
    "security": [
     {
      "OAuth2PasswordBearer": []
     }
    ],
    "parameters": [
     {
      "name": "building_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Building Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Space Board Api Space Spaces Board  Building Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "OAuth2PasswordBearer": {
    "type": "oauth2",
    "flows": {
     "password": {
      "scopes": {},
      "tokenUrl": "/api/auth/login"
     }
    }
   }
  }
 }
}